Separate openDRS session logic from fetching work order details
