+++ src/components/WorkOrder/WorkOrderDetails.js
@@ -13,8 +13,6 @@ const WorkOrderDetails = ({
   property,
   workOrder,
   tenure,
-  schedulerSessionId,
-  resetSchedulerSessionId,
   printClickHandler,
   setLocationAlerts,
   setPersonAlerts,
@@ -79,8 +77,6 @@ const WorkOrderDetails = ({
           tenure={tenure}
           hasLinkToProperty={true}
           canRaiseRepair={property.canRaiseRepair}
-          schedulerSessionId={schedulerSessionId}
-          resetSchedulerSessionId={resetSchedulerSessionId}
           setLocationAlerts={setLocationAlerts}
           setPersonAlerts={setPersonAlerts}
         />
