+++ src/components/WorkOrder/WorkOrderView.js
@@ -8,7 +8,6 @@ import Tabs from '../Tabs'
 import { WorkOrder } from '@/models/workOrder'
 import { sortObjectsByDateKey } from '@/utils/date'
 import PrintJobTicketDetails from './PrintJobTicketDetails'
-import { getOrCreateSchedulerSessionId } from '../../utils/frontEndApiClient/users/schedulerSession'
 
 const WorkOrderView = ({ workOrderReference }) => {
   const [property, setProperty] = useState({})
@@ -17,7 +16,6 @@ const WorkOrderView = ({ workOrderReference }) => {
   const [locationAlerts, setLocationAlerts] = useState([])
   const [personAlerts, setPersonAlerts] = useState([])
   const [tenure, setTenure] = useState({})
-  const [schedulerSessionId, setSchedulerSessionId] = useState()
   const [loading, setLoading] = useState(false)
   const [error, setError] = useState()
   const tabsList = [
@@ -27,11 +25,6 @@ const WorkOrderView = ({ workOrderReference }) => {
     'Work orders history',
     'Photos',
   ]
-
-  const getSchedulerSessionId = async () => {
-    const schedulerSessionId = await getOrCreateSchedulerSessionId()
-    setSchedulerSessionId(schedulerSessionId)
-  }
 
   const { NEXT_PUBLIC_STATIC_IMAGES_BUCKET_URL } = process.env
 
@@ -83,11 +76,6 @@ const WorkOrderView = ({ workOrderReference }) => {
         sortObjectsByDateKey(tasksAndSors, ['dateAdded'], 'dateAdded')
       )
 
-      // Call getOrCreateSchedulerSessionId if it is a DRS work order
-      if (workOrder.externalAppointmentManagementUrl) {
-        await getSchedulerSessionId()
-      }
-
       setWorkOrder(new WorkOrder(workOrder))
       setProperty(propertyObject.property)
       if (propertyObject.tenure) setTenure(propertyObject.tenure)
@@ -118,50 +106,44 @@ const WorkOrderView = ({ workOrderReference }) => {
     getWorkOrderView(workOrderReference)
   }, [])
 
+  if (loading) return <Spinner />
+
   return (
     <>
-      {loading ? (
-        <Spinner />
-      ) : (
-        <>
-          {property &&
-            property.address &&
-            property.hierarchyType &&
-            tenure &&
-            workOrder && (
-              <>
-                <WorkOrderDetails
-                  property={property}
-                  workOrder={workOrder}
-                  tenure={tenure}
-                  resetSchedulerSessionId={getSchedulerSessionId}
-                  schedulerSessionId={schedulerSessionId}
-                  tasksAndSors={tasksAndSors}
-                  printClickHandler={printClickHandler}
-                  setLocationAlerts={setLocationAlerts}
-                  setPersonAlerts={setPersonAlerts}
-                />
-                <Tabs
-                  tabsList={tabsList}
-                  propertyReference={property.propertyReference}
-                  workOrderReference={workOrderReference}
-                  tasksAndSors={tasksAndSors}
-                  budgetCode={workOrder.budgetCode}
-                  workOrder={workOrder}
-                />
-                {/* Only displayed for print media */}
-                <PrintJobTicketDetails
-                  workOrder={workOrder}
-                  property={property}
-                  tasksAndSors={tasksAndSors}
-                  locationAlerts={locationAlerts}
-                  personAlerts={personAlerts}
-                />
-              </>
-            )}
-          {error && <ErrorMessage label={error} />}
-        </>
-      )}
+      {property &&
+        property.address &&
+        property.hierarchyType &&
+        tenure &&
+        workOrder && (
+          <>
+            <WorkOrderDetails
+              property={property}
+              workOrder={workOrder}
+              tenure={tenure}
+              tasksAndSors={tasksAndSors}
+              printClickHandler={printClickHandler}
+              setLocationAlerts={setLocationAlerts}
+              setPersonAlerts={setPersonAlerts}
+            />
+            <Tabs
+              tabsList={tabsList}
+              propertyReference={property.propertyReference}
+              workOrderReference={workOrderReference}
+              tasksAndSors={tasksAndSors}
+              budgetCode={workOrder.budgetCode}
+              workOrder={workOrder}
+            />
+            {/* Only displayed for print media */}
+            <PrintJobTicketDetails
+              workOrder={workOrder}
+              property={property}
+              tasksAndSors={tasksAndSors}
+              locationAlerts={locationAlerts}
+              personAlerts={personAlerts}
+            />
+          </>
+        )}
+      {error && <ErrorMessage label={error} />}
     </>
   )
 }
